backend/pyproject.toml
[project]
name = "navi-backend"
version = "0.1.0"
description = "Navi — AI-Powered Stock Analyzer Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "google-cloud-firestore>=2.19.0",
    "firebase-admin>=6.5.0",
    "google-cloud-aiplatform>=1.60.0",
    "yfinance>=0.2.40",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
]

[project.scripts]
dev = "uvicorn main:app --reload --port 8000"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
